Refactor: Eliminate code duplication and improve architecture
Major architectural improvements to enhance code quality and maintainability:

1. Create BaseSimulator abstraction (eliminates ~200 lines of duplication)
   - Add core/simulators/base_simulator.py with shared simulation logic
   - Extract duplicate code from FishTankSimulator and SimulationEngine
   - Implement template method pattern for collision handling, reproduction, etc.
   - Both simulators now extend BaseSimulator with mode-specific implementations

2. Extract UI rendering into dedicated UIRenderer class
   - Add rendering/ui_renderer.py for all UI components
   - Separate rendering logic from simulation logic (separation of concerns)
   - Move draw_health_bar, draw_stats_panel, draw_poker_notifications to UIRenderer
   - Improve testability and reusability of UI components

3. Add comprehensive architecture documentation
   - Create ARCHITECTURE.md documenting design decisions
   - Document design patterns used (Template Method, Strategy, Dependency Injection)
   - Include before/after metrics showing code quality improvements
   - Provide future improvement recommendations

Benefits:
- Reduced code duplication by ~200 lines
- Improved separation of concerns
- Better maintainability and testability
- Consistent behavior between graphical and headless modes
- Clear architectural patterns for future development

Files changed:
- Created: core/simulators/__init__.py
- Created: core/simulators/base_simulator.py (262 lines)
- Created: rendering/ui_renderer.py (200 lines)
- Created: ARCHITECTURE.md
- Modified: fishtank.py (simplified, uses BaseSimulator and UIRenderer)
- Modified: simulation_engine.py (simplified, uses BaseSimulator)
