+++ core/simulators/__init__.py
@@ -1,0 +1,5 @@
+"""Simulator base classes and implementations."""
+
+from .base_simulator import BaseSimulator
+
+__all__ = ['BaseSimulator']
